Add a separate basic Attachment type
