+++ service/src/main/kotlin/fi/espoo/evaka/attachment/Attachment.kt
@@ -1,0 +1,14 @@
+package fi.espoo.evaka.attachment
+
+import fi.espoo.evaka.shared.AttachmentId
+
+data class Attachment(
+    val id: AttachmentId,
+    val name: String,
+    val contentType: String,
+)
+
+enum class AttachmentType {
+    URGENCY,
+    EXTENDED_CARE
+}
