+++ service/src/main/kotlin/fi/espoo/evaka/attachment/AttachmentQueries.kt
@@ -4,8 +4,6 @@
 
 package fi.espoo.evaka.attachment
 
-import fi.espoo.evaka.application.Attachment
-import fi.espoo.evaka.application.AttachmentType
 import fi.espoo.evaka.shared.ApplicationId
 import fi.espoo.evaka.shared.AttachmentId
 import fi.espoo.evaka.shared.IncomeStatementId
@@ -67,12 +65,13 @@ fun Database.Transaction.insertAttachment(
 fun Database.Read.getAttachment(id: AttachmentId): Attachment? = this
     .createQuery(
         """
-        SELECT id, name, content_type, updated, received_at, type, uploaded_by_employee, uploaded_by_person
+        SELECT id, name, content_type
         FROM attachment
         WHERE id = :id
         """
     )
-    .bind("id", id).mapTo<Attachment>()
+    .bind("id", id)
+    .mapTo<Attachment>()
     .firstOrNull()
 
 fun Database.Read.isOwnAttachment(attachmentId: AttachmentId, user: AuthenticatedUser): Boolean {
